requirements.txt
apache-airflow==2.6.3
pandas==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
lxml==4.9.3
python-dotenv==1.0.0
